stream rework; funcs split; fixes per coms
