+++ psarev.alexandr/P4/checkMatrix.cpp
@@ -1,0 +1,14 @@
+#include "checkMatrix.hpp"
+#include <fstream>
+
+bool psarev::isTriMatrix(int* matrix, int rows, int cols)
+{
+  for (int i = 1; i < rows; i++) {
+    for (int k = 0; k < i; k++) {
+      if (matrix[i * cols + k] != 0) {
+        return false;
+      }
+    }
+  }
+  return true;
+}
