+++ psarev.alexandr/P4/checkMatrix.hpp
@@ -1,0 +1,8 @@
+#ifndef CHECK_MATRIX_HPP
+#define CHECK_MATRIX_HPP
+
+namespace psarev
+{
+  bool isTriMatrix(int* matrix, int rows, int cols);
+}
+#endif
