+++ psarev.alexandr/P4/inputMatrix.hpp
@@ -1,0 +1,10 @@
+#ifndef INPUT_MATRIX_HPP
+#define INPUT_MATRIX_HPP
+#include <iosfwd>
+
+namespace psarev
+{
+  size_t inputMatrix(std::istream& in, int* matrix, size_t cap);
+}
+
+#endif
